js -> tsx; tab -> modal
