+++ ymir/web/src/pages/project/diagnose/components/CurveView.tsx
