+++ ymir/web/src/pages/project/diagnose/components/DefaultStages.tsx
