+++ ymir/web/src/pages/project/diagnose/components/PRCurve.tsx
