--- ymir/web/src/pages/project/diagnose/components/view.js
@@ -1,8 +1,0 @@
-const View = (Component) => {
-  const Viewer = (props) => {
-    return <Component {...props} />
-  }
-  return Viewer
-}
-
-export default View
